Use public_name instead of email (PrintModule)
Using `public_name` as substitute for users `email` if email is not
present. Leaving rest of email occurences in module as it is; once
stored to `print_subscriptions` table, we do not want to search for
alternative (eg. public_name).

Refactoring use of email as user identification used through system.
In future it will be either email, phone number or unique nickname
filled in `users.public_name`.

remp/crm#1145