Fix invalid condition in print export generation if there's a time change
The condition used printExportDate as it was provided by the export command.
It wasn't in line with the date stored in the DB by engine, which throws away
time and always uses midnight of the export date column.

This caused the condition to be off 1 hour if there's a time change and the
records would not be updated with their states correctly.

remp/helpdesk#450